generate hash's prefix based on PHP version
With the release of PHP 5.3.7 crypt_blowfish was upgraded to version 1.2,
solving an issue with insecure passwords due to incorrect behaviour. See
http://www.openwall.com/lists/announce/2011/06/21/1 and
https://bugs.php.net/bug.php?id=55477 for more information.

php-bcrypt-wrapper will generate prefix based on PHP version, but it will NOT
try to fix "broken" hash automatically. If you provide hash generated with
pre-PHP 5.3.7, Bcrypt::check() will return false. In this case, you need to
replace $2a$ with $2x$, verify it and re-hash it again.